add frontend loader btn and backend validations and exceptions

+++ backend/src/main/java/com/f42o/api/transaction/MonthlyTransactionDTO.java
@@ -16,5 +16,5 @@ public class MonthlyTransactionDTO {
 
     private SummaryType type;
     private BigDecimal total;
-    private Double differenceBetweenPastMonth;
+    private BigDecimal differenceBetweenPastMonth;
 }
